+++ index.js
@@ -1,27 +1,8 @@
 require('dotenv').config();
-const {
-    SQSClient,
-    ReceiveMessageCommand,
-    DeleteMessageCommand,
-} = require('@aws-sdk/client-sqs');
 const examples = require('./example');
 const { Configuration, OpenAIApi } = require('openai');
-
-const sqsClient = new SQSClient({
-    region: process.env.AWS_REGION,
-    credentials: {
-        accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-    },
-});
-
-const queueURL = process.env.SQS_QUEUE_URL;
-
-const receiveMessageCommand = new ReceiveMessageCommand({
-    QueueUrl: queueURL,
-    MaxNumberOfMessages: 1,
-    WaitTimeSeconds: 20,
-});
+const { getMessage, deleteMessage } = require('./queue');
+const { isValidResponse } = require('./validation');
 
 const configuration = new Configuration({
     apiKey: process.env.AI_API_KEY,
@@ -131,46 +112,6 @@ const generate = async (inputText, prompt) => {
 
 const input = examples.example2;
 
-const isValidCard = (card) => {
-    return (
-        typeof card === 'object' &&
-        card.front &&
-        typeof card.front === 'string' &&
-        card.back &&
-        typeof card.back === 'string'
-    );
-};
-
-const isValidResponse = (response, expectedCardCount) => {
-    if (typeof response !== 'object' || !response.flashcards) {
-        console.error('Invalid response');
-        return false;
-    }
-    const { flashcards } = response;
-
-    if (!Array.isArray(flashcards)) {
-        console.error('Flashcards is not an array');
-        return false;
-    }
-
-    if (flashcards.length < expectedCardCount) {
-        console.error('Less cards returned than expected', flashcards.length);
-        return false;
-    }
-    if (flashcards.length > expectedCardCount) {
-        console.error('More cards returned than expected', flashcards.length);
-        return false;
-    }
-
-    for (let card of flashcards) {
-        if (!isValidCard(card)) {
-            console.error('Invalid card:', card);
-            return false;
-        }
-    }
-    return true;
-};
-
 const runTest = async (input, prompt) => {
     const results = {
         failed: 0,
@@ -187,7 +128,7 @@ const runTest = async (input, prompt) => {
 
         console.log(i);
         console.log(json);
-        const result = isValidResponse(json, 1);
+        const result = isValidResponse(json);
         // add logic to get full count of failed and successes
         // as well as what the responses were that failed
         if (!result) {
@@ -198,8 +139,11 @@ const runTest = async (input, prompt) => {
     console.log(results);
 };
 
+//runTest(input, promptBuilder({ cardCount: 5 }));
+
 const generateFromQueue = async () => {
-    const data = await sqsClient.send(receiveMessageCommand);
+    // TODO validate response
+    const data = await getMessage();
     console.log('called');
     console.log(data);
     if (!data.Messages) {
@@ -207,18 +151,39 @@ const generateFromQueue = async () => {
         generateFromQueue();
         return;
     }
-    const message = JSON.parse(data.Messages[0].Body);
+
+    let message;
+    try {
+        message = JSON.parse(data.Messages[0].Body);
+    } catch (err) {
+        console.error(err);
+        generateFromQueue();
+        return;
+    }
+
     const prompt = promptBuilder(message);
     const flashcards = await generate(message.inputText, prompt);
     console.log(flashcards);
-    const deleteMessageCommand = new DeleteMessageCommand({
-        QueueUrl: queueURL,
-        ReceiptHandle: data.Messages[0].ReceiptHandle,
-    });
-    const resp = await sqsClient.send(deleteMessageCommand);
+    const res = await deleteMessage(data.Messages[0].ReceiptHandle);
+    if (res.$metadata.httpStatusCode !== 200) {
+        // TODO add retry logic
+    }
+    console.log(res);
     generateFromQueue();
 };
+
 generateFromQueue();
 
-// TODO remove messages from queue after processing
+// TODO refactor
+// TODO add validation
 // TODO set up proper error handling
+// TODO add retry logic
+// TODO set up proper logging
+// TODO Add rate limiting
+
+// Failure points:
+// 1. Invalid response from API
+// 2. Invalid response from queue
+// 3. Invalid response from validation
+// 4. Invalid response from queue deletion
+// 5. Invalid response from queue message retrieval
