+++ package.json
@@ -5,6 +5,7 @@
     "main": "index.js",
     "scripts": {
         "start": "node index.js",
+        "enqueue": "node enqueue.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "",
